fix TUs

+++ Pipeline/Model/FileImage.py
@@ -5,16 +5,17 @@ import matplotlib.gridspec as gridspec
 from scipy.misc import imsave
 
 class FileImage:
-    image: []
 
     def __init__(self, fullname: str = None):
         self.log = logging.getLogger('IMG')
         self.local = pytz.timezone("Europe/Paris")
+        self.image = []
         self.UpdateFullName(fullname)
 
     def UpdateFullName(self, fullname):
         if fullname == None:
-            return        
+            return       
+        self.EnsureImage() 
         self.fullname = fullname
         self.filename = os.path.basename(fullname)
         self.dir = os.path.dirname(fullname)
@@ -31,8 +32,11 @@ class FileImage:
         fp.write(content)
         fp.close()
 
-    def Save(self):
+    def Save(self, fullname: str = None):
         params = list()
+        self.EnsureImage()
+        if fullname:
+            self.UpdateFullName(fullname)
         image = cv2.cvtColor(self.image, cv2.COLOR_RGB2BGR)
         if self.filenameExtension == 'png':
             imsave(self.fullname,image,params)
@@ -47,15 +51,22 @@ class FileImage:
         #self.log.info(f"Load image from: {self.fullname}")
         #self.image = mpimg.imread(self.fullname)
         self.image = cv2.imread(self.fullname, cv2.IMREAD_UNCHANGED)
+        #self.log.info(f"Image: {len(self.image)}")
         self.image = cv2.cvtColor(self.image, cv2.COLOR_BGR2RGB)
 
+    def EnsureImage(self):
+        if len(self.image) == 0 and hasattr(self, 'fullname') and self.fullname:
+            self.LoadImage()
+
+    def GetImage(self):
+        self.EnsureImage()
+        return self.image
 
     def Show(self):
         plt.figure(figsize=(8, 6.025))
         self.gs1 = gridspec.GridSpec(1, 1, left=0, right=1, top=1,
                                      bottom=0, wspace=0, hspace=0)
-        if len(self.image) == 0:
-            self.LoadImage()
+        self.EnsureImage()
         plt.subplot(self.gs1[0])
         plt.axis("off")
         plt.imshow(self.image, interpolation="bilinear")
@@ -67,8 +78,12 @@ class FileImage:
         self.UpdateFullName(dest)
         
     def Move2(self, dest: str):
-        orig = self.fullname
-        shutil.copy2(self.fullname, dest)
-        self.UpdateFullName(dest)
-        os.remove(orig)
+        if self.Exist():
+            orig = self.fullname
+            shutil.copy2(self.fullname, dest)
+            self.UpdateFullName(dest)
+            os.remove(orig)
+        elif len(self.image) != 0:
+            self.UpdateFullName(dest)
+            self.Save()
